requirements.txt
Django>=5.0,<5.1
Pillow>=10.0
weasyprint>=60.0
whitenoise>=6.0
psycopg2-binary>=2.9
gunicorn>=21.0
google-genai
pypdf
